Remove logic from old sdformat script
The sdformat-base-windows.bat file is only used
for sdformat5 and earlier, so we can remove logic related to sdformat's
default/master branch, and logic for ign-math4.
Also the IGNMATH_BRANCH variable is deduced from cmake/SearchForStuff
and is not read as a parameter, so remove it from the parameter
documentation.

Signed-off-by: Steven Peters <scpeters@openrobotics.org>