Keep more package build artifacts. (#524)
Package builders were recently updated (#521) to take advantage of an
increased number of jenkins agents and this has surfaced a bug which has
been dormant due to the lack of concurrent jobs.

Ignition Fortress releases currently have 13 platform combinations of
different OS versions and CPU architectures. Which means that with the
previous archiveToKeep value of 10, if the eleventh debbuilder job runs
before the first is copied to the packaging server then the artifacts
will be cleaned up before they can be imported.

The new value of 25 is more than double the current number of target
platforms but is sadly still a constant and not based on the number of
target platforms. I still think it is a more sensible default than 10
given the current situation.

With this change we will have to monitor the impact on the storage
capacity of the Jenkins server where these artifacts are stored.