Run mk-build-deps from a temporary directory.
Newer versions of mk-build-deps leave behind residual files. Other
scripts in this repository, notably
jenkins-scripts/docker/lib/debbuild-base.bash handle this by explicitly
removing the created files, but I am proposing that we handle it by
creating the package in a temporary directory and then cleaning the
entire directory. This way any changes in the future which, for example,
generate additional leftovers, will be handled transparently rather than
requiring an update.

If this approach is preferred, I'll update the other file as well.

If not I'll rework this to use the same mechanism as the existing
handler.