changelog script: improve tag/branch detection
Currently the source_changelog.bash script only
works for ignition-* tags, while garden will use
gz-* tags. This updates the script to grep for
a tag that ends with the specified version number.
It also uses the current branch for the comparison
to avoid assumptions about ign-* or gz-* branch
names.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>