ABI script: get correct header paths
Fixes #890. The ABI script currently assumes that
the headers are installed in a subfolder that
starts with the ABI_JOB_HEADER_PREFIX variable
and a -, which only works for sdformat-9 but
requires adjustment for most other projects.
This adds a ABI_JOB_HEADER_PREFIX variable to
help locate the header folder.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>