Add OSRFReleasepy DSL class
Wrapper to call release.py using Jenkins parameters as
release.py arguments. Note that the script is not fully
functional for standard releases since it currently requires
a local software checkout to be run.

This can work with nightly builds or future calls to
release.py that don't need to generate the tarballs.

The test inject DRY_RUN to avoid any real run.