Implement job chain from tarball to call releasepy and repo_uploader
Wrapper implementation to manage releasepy and repository_uploader
sequentially and reacting to failures. This is thanks to use the
downstream calls as build steps.

The tarball generation job can not be integrated since it needs to
use archiveArtifacts before calling other jobs and the artifacts are
archived in post build steps.