Rewrite DSL generation to remove the approach based on Indexes (#1137)
The DSL generation in gazebo_libs is based on an approach of building two indexes before looping over the gz-collections.yaml file. That approach does not respect the fact of having different ci_config (CI confgurations) that use the same platform and will overwrite the first one found if that situation occurs. The index build process also did not load the exclusion clauses included in the ci_configuration which are useful to be parsed when building the index.

The changes in this PR modify it to mainly:

 * Loop over the gz-collections.yaml directly the first time to generate the jobs that are linked to lib branches gz-fooX.
 * While looping it build a branch_index that will be used to build the pr jobs that are in the form gz-foo-pr_any- not based on the branch directly.
  * Loop over the branch_index created to create the pr jobs.

Note: the common jobs used for -pr- need the assumption of having the same metadata when ci_configs use the same platform. The restrictions needs to be coded or the design corrected with structural changes.

--------
Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>